Rename ActionableDataTable to DataGrid
