+++ src/components/index.ts
@@ -1,4 +1,4 @@
-export { default as ActionableDataTable } from './ActionableDataTable.vue';
+export { default as DataGrid } from './DataGrid.vue';
 export { default as Attendance } from './Attendance.vue';
 export { default as Card } from './card.vue';
 export { default as DualListbox } from './dual-listbox.vue';
